@@ -1,16 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net6.0</TargetFramework>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="3.10.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\AnyOfCodeGenerator\AnyOf.SourceGenerator.csproj" />
-  </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\AnyOfCodeGenerator\AnyOf.SourceGenerator.csproj" />
+    </ItemGroup>
 
 </Project>
