@@ -5,104 +5,118 @@ using System.Linq;
 using System.Reflection;
 using AnyOfTypes.System.Text.Json.Matcher.Models;
 
-namespace AnyOfTypes.System.Text.Json.Extensions
+namespace AnyOfTypes.System.Text.Json.Extensions;
+
+internal static class ReflectionHelpers
 {
-    internal static class ReflectionHelpers
+    [ThreadStatic]
+    static readonly Dictionary<KeyValuePair<Type, Type>, bool> ImplicitCastCache;
+
+    static ReflectionHelpers()
     {
-        [ThreadStatic]
-        static readonly Dictionary<KeyValuePair<Type, Type>, bool> ImplicitCastCache = new Dictionary<KeyValuePair<Type, Type>, bool>();
+        ImplicitCastCache = new Dictionary<KeyValuePair<Type, Type>, bool>();
+    }
 
-        public static T GetPropertyValue<T>(this object instance, string name)
+    public static bool IsNullableType(this Type type)
+    {
+        return type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>);
+    }
+
+    public static Type GetNonNullableType(this Type type)
+    {
+        return IsNullableType(type) ? type.GetTypeInfo().GenericTypeArguments[0] : type;
+    }
+
+    public static T GetPropertyValue<T>(this object instance, string name)
+    {
+        var value = GetNullablePropertyValue(instance, name);
+        if (value is null)
         {
-            var value = GetNullablePropertyValue(instance, name);
-            if (value is null)
-            {
-                throw new InvalidOperationException($"The public property '{name}' has a null value.");
-            }
-
-            return (T)value;
+            throw new InvalidOperationException($"The public property '{name}' has a null value.");
         }
 
-        public static object? GetNullablePropertyValue(this object instance, string name)
-        {
-            var type = instance.GetType();
-            var propertyInfo = type.GetProperty(name);
-            if (propertyInfo is null)
-            {
-                throw new InvalidOperationException($"The type '{type}' does not contain public property '{name}'.");
-            }
+        return (T)value;
+    }
 
-            return propertyInfo.GetValue(instance);
+    public static object? GetNullablePropertyValue(this object instance, string name)
+    {
+        var type = instance.GetType();
+        var propertyInfo = type.GetProperty(name);
+        if (propertyInfo is null)
+        {
+            throw new InvalidOperationException($"The type '{type}' does not contain public property '{name}'.");
         }
 
-        public static Type GetElementTypeX(this Type enumerableType)
+        return propertyInfo.GetValue(instance);
+    }
+
+    public static Type GetElementTypeX(this Type enumerableType)
+    {
+        return enumerableType.IsArray ? enumerableType.GetElementType()! : enumerableType.GetGenericArguments().First();
+    }
+
+    public static ListDetails CastToTypedList(this IList source, Type elementType)
+    {
+        var listType = typeof(List<>).MakeGenericType(elementType);
+        var list = (IList)Activator.CreateInstance(listType);
+
+        foreach (var item in source)
         {
-            return enumerableType.IsArray == true ? enumerableType.GetElementType() : enumerableType.GetGenericArguments().First();
+            list.Add(item);
         }
 
-        public static ListDetails CastToTypedList(this IList source, Type elementType)
+        return new ListDetails
         {
-            var listType = typeof(List<>).MakeGenericType(elementType);
-            var list = (IList)Activator.CreateInstance(listType);
-            
-            foreach (var item in source)
-            {
-                list.Add(item);
-            }
+            List = list,
+            ListType = listType
+        };
+    }
 
-            return new ListDetails
-            {
-                List = list,
-                ListType = listType
-            };
+    /// <summary>
+    /// https://stackoverflow.com/questions/17676838/how-to-check-if-type-can-be-converted-to-another-type-in-c-sharp
+    /// https://stackoverflow.com/questions/2224266/how-to-tell-if-type-a-is-implicitly-convertible-to-type-b
+    /// </summary>
+    public static bool IsImplicitlyCastableTo(this Type? from, Type? to)
+    {
+        if (from is null || to is null)
+        {
+            return false;
         }
 
-        /// <summary>
-        /// https://stackoverflow.com/questions/17676838/how-to-check-if-type-can-be-converted-to-another-type-in-c-sharp
-        /// https://stackoverflow.com/questions/2224266/how-to-tell-if-type-a-is-implicitly-convertible-to-type-b
-        /// </summary>
-        public static bool IsImplicitlyCastableTo(this Type? from, Type? to)
+        if (from == to)
         {
-            if (from is null || to is null)
-            {
-                return false;
-            }
+            return true;
+        }
 
-            if (from == to)
-            {
-                return true;
-            }
-
-            var key = new KeyValuePair<Type, Type>(from, to);
-            if (ImplicitCastCache.TryGetValue(key, out bool result))
-            {
-                return result;
-            }
+        var key = new KeyValuePair<Type, Type>(from, to);
+        if (ImplicitCastCache.TryGetValue(key, out bool result))
+        {
+            return result;
+        }
 
 #if !NETSTANDARD1_3
-            if (to.IsAssignableFrom(from))
-            {
-                return ImplicitCastCache[key] = true;
-            }
-#endif
-            if (from.GetMethods(BindingFlags.Public | BindingFlags.Static).Any(m => m.ReturnType == to && (m.Name == "op_Implicit" || m.Name == "op_Explicit")))
-            {
-                return ImplicitCastCache[key] = true;
-            }
-
-            bool changeType = false;
-            try
-            {
-                var val = Activator.CreateInstance(from);
-                Convert.ChangeType(val, to);
-                changeType = true;
-            }
-            catch
-            {
-                // Ignore
-            }
-
-            return ImplicitCastCache[key] = changeType;
+        if (to.IsAssignableFrom(from))
+        {
+            return ImplicitCastCache[key] = true;
         }
+#endif
+        if (from.GetMethods(BindingFlags.Public | BindingFlags.Static).Any(m => m.ReturnType == to && (m.Name == "op_Implicit" || m.Name == "op_Explicit")))
+        {
+            return ImplicitCastCache[key] = true;
+        }
+
+        bool changeType = false;
+        try
+        {
+            var val = Activator.CreateInstance(from);
+            _ = Convert.ChangeType(val, to);
+            changeType = true;
+        }
+        catch
+        {
+            // Ignore
+        }
+
+        return ImplicitCastCache[key] = changeType;
     }
 }
