@@ -1,9 +1,8 @@
-﻿namespace AnyOfTypes.System.Text.Json.Matcher.Models
-{
-    internal struct PropertyMap
-    {
-        public PropertyDetails SourceProperty { get; set; }
+﻿namespace AnyOfTypes.System.Text.Json.Matcher.Models;
 
-        public PropertyDetails TargetProperty { get; set; }
-    }
+internal struct PropertyMap
+{
+    public PropertyDetails SourceProperty { get; set; }
+
+    public PropertyDetails TargetProperty { get; set; }
 }
