@@ -4,237 +4,239 @@ using FluentAssertions;
 using Newtonsoft.Json;
 using Xunit;
 
-namespace AnyOfTypes.Newtonsoft.Json.Tests
+// ReSharper disable once CheckNamespace
+namespace AnyOfTypes.Newtonsoft.Json.Tests;
+
+public class AnyOfJsonConverterTests
 {
-    public class AnyOfJsonConverterTests
+    [Fact]
+    public void Serialize_AnyOf_With_SimpleTypes()
     {
-        [Fact]
-        public void Serialize_AnyOf_With_SimpleTypes()
+        // Arrange
+        var test = new TestSimpleTypes
         {
-            // Arrange
-            var test = new TestSimpleTypes
-            {
-                IntOrString = 1
-            };
+            IntOrString = 1,
+            NullableIntOrString = "s"
+        };
 
-            // Act
-            var options = new JsonSerializerSettings
-            {
-                Formatting = Formatting.None
-            };
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var json = JsonConvert.SerializeObject(test, options);
-
-            // Assert
-            json.Should().Be("{\"IntOrString\":1}");
-        }
-
-        [Fact]
-        public void Serialize_AnyOf_With_ComplexTypes()
+        // Act
+        var options = new JsonSerializerSettings
         {
-            // Arrange
-            var test = new TestComplexTypes
-            {
-                AorB = new A
-                {
-                    Id = 1
-                }
-            };
+            Formatting = Formatting.None
+        };
+        options.Converters.Add(new AnyOfJsonConverter());
 
-            // Act
-            var options = new JsonSerializerSettings
-            {
-                Formatting = Formatting.None
-            };
-            options.Converters.Add(new AnyOfJsonConverter());
+        var json = JsonConvert.SerializeObject(test, options);
 
-            var json = JsonConvert.SerializeObject(test, options);
+        // Assert
+        json.Should().Be("{\"IntOrString\":1,\"NullableIntOrString\":\"s\"}");
+    }
 
-            // Assert
-            json.Should().Be("{\"AorB\":{\"Id\":1}}");
-        }
-
-        [Fact]
-        public void Serialize_AnyOf_With_MixedTypes()
+    [Fact]
+    public void Serialize_AnyOf_With_ComplexTypes()
+    {
+        // Arrange
+        var test = new TestComplexTypes
         {
-            // Arrange
-            var test = new TestMixedTypes
-            {
-                IntOrStringOrAOrB = 1
-            };
-
-            // Act
-            var options = new JsonSerializerSettings
-            {
-                Formatting = Formatting.None
-            };
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var json = JsonConvert.SerializeObject(test, options);
-
-            // Assert
-            json.Should().Be("{\"IntOrStringOrAOrB\":1}");
-        }
-
-        [Fact]
-        public void Deserialize_AnyOf_With_SimpleTypes()
-        {
-            // Arrange
-            var expected = new TestSimpleTypes
-            {
-                IntOrString = 1
-            };
-
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var result = JsonConvert.DeserializeObject<TestSimpleTypes>("{\"IntOrString\":1}", options);
-
-            // Assert
-            result.Should().BeEquivalentTo(expected);
-        }
-
-        [Fact]
-        public void Deserialize_AnyOf_With_ComplexTypes()
-        {
-            // Arrange
-            var expected = new A
+            AorB = new A
             {
                 Id = 1
-            };
+            }
+        };
 
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var result = JsonConvert.DeserializeObject<TestComplexTypes>("{\"AorB\":{\"Id\":1}}", options);
-
-            // Assert
-            result.AorB.First.Should().BeEquivalentTo(expected);
-        }
-
-        [Fact]
-        public void Deserialize_AnyOf_With_ComplexTypes_DifferentCasing()
+        // Act
+        var options = new JsonSerializerSettings
         {
-            // Arrange
-            var expected = new A2
+            Formatting = Formatting.None
+        };
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var json = JsonConvert.SerializeObject(test, options);
+
+        // Assert
+        json.Should().Be("{\"AorB\":{\"Id\":1}}");
+    }
+
+    [Fact]
+    public void Serialize_AnyOf_With_MixedTypes()
+    {
+        // Arrange
+        var test = new TestMixedTypes
+        {
+            IntOrStringOrAOrB = 1
+        };
+
+        // Act
+        var options = new JsonSerializerSettings
+        {
+            Formatting = Formatting.None
+        };
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var json = JsonConvert.SerializeObject(test, options);
+
+        // Assert
+        json.Should().Be("{\"IntOrStringOrAOrB\":1}");
+    }
+
+    [Fact]
+    public void Deserialize_AnyOf_With_SimpleTypes()
+    {
+        // Arrange
+        var expected = new TestSimpleTypes
+        {
+            IntOrString = 1,
+            NullableIntOrString = "s"
+        };
+
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var result = JsonConvert.DeserializeObject<TestSimpleTypes>("{\"IntOrString\":1,\"NullableIntOrString\":\"s\"}", options);
+
+        // Assert
+        result.Should().BeEquivalentTo(expected);
+    }
+
+    [Fact]
+    public void Deserialize_AnyOf_With_ComplexTypes()
+    {
+        // Arrange
+        var expected = new A
+        {
+            Id = 1
+        };
+
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var result = JsonConvert.DeserializeObject<TestComplexTypes>("{\"AorB\":{\"Id\":1}}", options)!;
+
+        // Assert
+        result.AorB.First.Should().BeEquivalentTo(expected);
+    }
+
+    [Fact]
+    public void Deserialize_AnyOf_With_ComplexTypes_DifferentCasing()
+    {
+        // Arrange
+        var expected = new A2
+        {
+            id = 1
+        };
+
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var result = JsonConvert.DeserializeObject<TestComplexTypes2>("{\"AorB\":{\"Id\":1}}", options)!;
+
+        // Assert
+        result.AorB.First.Should().BeEquivalentTo(expected);
+    }
+
+    [Fact]
+    public void Deserialize_AnyOf_With_MixedTypes()
+    {
+        // Arrange
+        var expected = new TestMixedTypes
+        {
+            IntOrStringOrAOrB = 1
+        };
+
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var result = JsonConvert.DeserializeObject<TestMixedTypes>("{\"IntOrStringOrAOrB\":1}", options);
+
+        // Assert
+        result.Should().BeEquivalentTo(expected);
+    }
+
+    [Fact]
+    public void Deserialize_AnyOf_With_IntArray()
+    {
+        // Arrange
+        var expected = new[] { 42 };
+
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[42]}", options)!;
+
+        // Assert
+        result.X.First.Should().BeEquivalentTo(expected);
+    }
+
+    [Fact]
+    public void Deserialize_AnyOf_With_StringList()
+    {
+        // Arrange
+        var expected = new[] { "a", "b" };
+
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
+
+        var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[\"a\", \"b\"]}", options)!;
+
+        // Assert
+        result.X.Second.Should().BeEquivalentTo(expected);
+    }
+
+    [Fact]
+    public void Deserialize_AnyOf_With_ObjectList_A()
+    {
+        // Arrange
+        var expected = new List<A>
+        {
+            new A
             {
-                id = 1
-            };
-
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var result = JsonConvert.DeserializeObject<TestComplexTypes2>("{\"AorB\":{\"Id\":1}}", options);
-
-            // Assert
-            result.AorB.First.Should().BeEquivalentTo(expected);
-        }
-
-        [Fact]
-        public void Deserialize_AnyOf_With_MixedTypes()
-        {
-            // Arrange
-            var expected = new TestMixedTypes
+                Id = 1
+            },
+            new A
             {
-                IntOrStringOrAOrB = 1
-            };
+                Id = 2
+            }
+        };
 
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
 
-            var result = JsonConvert.DeserializeObject<TestMixedTypes>("{\"IntOrStringOrAOrB\":1}", options);
+        var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[{\"Id\":1},{\"Id\":2}]}", options)!;
 
-            // Assert
-            result.Should().BeEquivalentTo(expected);
-        }
+        // Assert
+        result.X.Third.Should().BeEquivalentTo(expected);
+    }
 
-        [Fact]
-        public void Deserialize_AnyOf_With_IntArray()
+    [Fact]
+    public void Deserialize_AnyOf_With_ObjectList_B()
+    {
+        // Arrange
+        var expected = new List<B>
         {
-            // Arrange
-            var expected = new int[] { 42 };
-
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[42]}", options);
-
-            // Assert
-            result.X.First.Should().BeEquivalentTo(expected);
-        }
-
-        [Fact]
-        public void Deserialize_AnyOf_With_StringList()
-        {
-            // Arrange
-            var expected = new[] { "a", "b" };
-
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[\"a\", \"b\"]}", options);
-
-            // Assert
-            result.X.Second.Should().BeEquivalentTo(expected);
-        }
-
-        [Fact]
-        public void Deserialize_AnyOf_With_ObjectList_A()
-        {
-            // Arrange
-            var expected = new List<A>
+            new B
             {
-                new A
-                {
-                    Id = 1
-                },
-                new A
-                {
-                    Id = 2
-                }
-            };
-
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
-
-            var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[{\"Id\":1},{\"Id\":2}]}", options);
-
-            // Assert
-            result.X.Third.Should().BeEquivalentTo(expected);
-        }
-
-        [Fact]
-        public void Deserialize_AnyOf_With_ObjectList_B()
-        {
-            // Arrange
-            var expected = new List<B>
+                Guid = "a"
+            },
+            new B
             {
-                new B
-                {
-                    Guid = "a"
-                },
-                new B
-                {
-                    Guid = "b"
-                }
-            };
+                Guid = "b"
+            }
+        };
 
-            // Act
-            var options = new JsonSerializerSettings();
-            options.Converters.Add(new AnyOfJsonConverter());
+        // Act
+        var options = new JsonSerializerSettings();
+        options.Converters.Add(new AnyOfJsonConverter());
 
-            var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[{\"Guid\":\"a\"},{\"Guid\":\"b\"}]}", options);
+        var result = JsonConvert.DeserializeObject<TestComplexArray>("{\"X\":[{\"Guid\":\"a\"},{\"Guid\":\"b\"}]}", options)!;
 
-            // Assert
-            result.X.Fourth.Should().BeEquivalentTo(expected);
-        }
+        // Assert
+        result.X.Fourth.Should().BeEquivalentTo(expected);
     }
 }
